+++ lib/src/flutter/platforms.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 
 /// Sample
 ///
-/// class PlatformAppBar extends VxPlatform<AppBar, CupertinoNavigationBar, AppBar> {
+/// class PlatformAppBar extends VxPlatform<AppBar, CupertinoNavigationBar, AppBar, AppBar> {
 ///   final Widget title;
 ///   final Widget leading;
 ///   PlatformAppBar({
@@ -14,26 +14,31 @@ import 'package:flutter/material.dart';
 ///
 ///   @override
 ///   CupertinoNavigationBar createIosWidget(BuildContext context) =>
-///       new CupertinoNavigationBar(
+///        CupertinoNavigationBar(
 ///         leading: leading,
 ///         middle: title,
 ///       );
 ///
 ///   @override
-///   AppBar createAndroidWidget(BuildContext context) => new AppBar(
+///   AppBar createAndroidWidget(BuildContext context) =>  AppBar(
 ///         leading: leading,
 ///         title: title,
 ///       );
 ///   @override
-///   AppBar createWebWidget(BuildContext context) => new AppBar(
+///   AppBar createWebWidget(BuildContext context) =>  AppBar(
+///         leading: leading,
+///         title: title,
+///       );
+///   @override
+///   AppBar createDesktopWidget(BuildContext context) =>  AppBar(
 ///         leading: leading,
 ///         title: title,
 ///       );
 ///}
 
 ///
-abstract class VxPlatform<A extends Widget, I extends Widget, W extends Widget>
-    extends StatelessWidget {
+abstract class VxPlatform<A extends Widget, I extends Widget, W extends Widget,
+    D extends Widget> extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     if (kIsWeb) {
@@ -42,6 +47,10 @@ abstract class VxPlatform<A extends Widget, I extends Widget, W extends Widget>
       return createAndroidWidget(context);
     } else if (defaultTargetPlatform == TargetPlatform.iOS) {
       return createIosWidget(context);
+    } else if (defaultTargetPlatform == TargetPlatform.macOS ||
+        defaultTargetPlatform == TargetPlatform.linux ||
+        defaultTargetPlatform == TargetPlatform.windows) {
+      return createDesktopWidget(context);
     }
     // platform not supported returns an empty widget
     return const Offstage();
@@ -50,4 +59,5 @@ abstract class VxPlatform<A extends Widget, I extends Widget, W extends Widget>
   A createAndroidWidget(BuildContext context);
   I createIosWidget(BuildContext context);
   W createWebWidget(BuildContext context);
+  D createDesktopWidget(BuildContext context);
 }